rearrange. hide std::map, resurrect destruction
Another type, D_DISPATCHER, is required for non static dispatchers
(which use is questionable). Could as well use DISPATCHER_BASE for this
purpose. with this patch, DISPATCHER_BASE is a class with ordinary
construction and destruction. DISPATCHER is no longer derived from it,
but deals with types, and manages a pointer to make the static use work.

This in effect changes
std::map<..>* DISPATCHER::_map;
to
DISPATCHER_BASE<..>* DISPATCHER::_base;
and std::map becomes an implementation detail of DISPATCHER_BASE.

Side effect, DISPATCHER_BASE is destroyed during shutdown, as it was in
the past. In debug mode, it checks if everything has been uninstalled
correctly (good).

NB: I don't know the motivation for the DISPATCHER/DISPATCHER_BASE
split. But this makes some use of that split, but breaks "is-a".

TODO:
- drop uninstall(string), not needed.
- drop D_DISPATCHER, could be done elsewhere, or use _BASE.