virtual-portlist: make nodelist access ( n_(i) ) virtual, and use it.
This commit makes the primary access to a node list ( n_(i) ) virtual,
and eliminates the old C-style _n pointer to the real node list.
All use of the _n pointer has been replaced by the access function,
replacing _n[i] with n_(i).

Any class having a node list must define (or inherit) the access function
n_(i), usually simply as {return _nodes[i];} or something like that.

Fast access is important, so bounds checking is only by assert().
As always, the node list is not just ports.  Internal nodes (nets)
are here too.

The root of n_(i) is still in CARD, because it is called that way,
for anything that could be in a CARDLIST, even if not appropriate.