predictor-1: updates to predictor 	Do not use if dtdt > 100 (or as set by OPT::predictor
The predictor was overshooting by extreme amounts when a small step
is followed by a large step, so revert to non-predictor when step
size increses by more than 100x.

The most common place it came up was a logic node changing state,
leading to a small step with a large change, followed by a
back-to-normal large step.  An option sets the threshold to 100x.
Step ratios smaller than this use the predictor (Forward Euler),
larger uses a repeat of the old value, as it was always pre-predictor.

Either way, the normal iteration gives a correct result.  Significant
differences are only in the 0th iteration of a new step. Regressions
show many insignificant changes as iteration runs a little different,
usually with fewer iterations required, although this is not the
purpose.  The purpose is to fill in latent nodes.